backend/requirements.txt
fastapi
uvicorn
langchain
openai
python-dotenv
langgraph
langchain-google-genai